split up scripts st register script so registering the bot can be run locally
